+++ app/Models/Product.php
@@ -5,6 +5,7 @@ namespace App\Models;
 use Database\Factories\ProductFactory;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
+use Illuminate\Database\Eloquent\Relations\BelongsToMany;
 
 /**
  * Модель продукта
@@ -20,4 +21,14 @@ class Product extends Model
     protected $fillable = [
         'title',
     ];
+
+    /**
+     * Получить рецепты, в которых используется текущий продукт
+     *
+     * @return BelongsToMany
+     */
+    public function recipes(): BelongsToMany
+    {
+        return $this->belongsToMany(Recipe::class);
+    }
 }