+++ database/migrations/2026_07_30_112714_create_recipes_table.php
@@ -1,0 +1,37 @@
+<?php
+
+use Illuminate\Database\Migrations\Migration;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Facades\Schema;
+
+/**
+ * Миграция таблицы рецептов
+ */
+return new class extends Migration
+{
+    /**
+     * Run the migrations.
+     */
+    public function up(): void
+    {
+        Schema::create('recipes', function (Blueprint $table) {
+            $table->id();
+            $table->timestamps();
+            $table->string('title', 255)
+                ->nullable(false)
+                ->default('')
+                ->charset('utf8mb4')
+                ->collation('utf8mb4_unicode_ci')
+                ->comment('название рецепта');
+            $table->unique('title');
+        });
+    }
+
+    /**
+     * Reverse the migrations.
+     */
+    public function down(): void
+    {
+        Schema::dropIfExists('recipes');
+    }
+};