+++ database/migrations/2026_07_30_120749_create_product_recipe_table.php
@@ -1,0 +1,31 @@
+<?php
+
+use App\Models\Product;
+use App\Models\Recipe;
+use Illuminate\Database\Migrations\Migration;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Facades\Schema;
+
+return new class extends Migration
+{
+    /**
+     * Run the migrations.
+     */
+    public function up(): void
+    {
+        Schema::create('product_recipe', function (Blueprint $table) {
+            $table->foreignIdFor(Product::class)->constrained()->cascadeOnDelete();
+            $table->foreignIdFor(Recipe::class)->constrained()->cascadeOnDelete();
+
+            $table->primary(['product_id', 'recipe_id']);
+        });
+    }
+
+    /**
+     * Reverse the migrations.
+     */
+    public function down(): void
+    {
+        Schema::dropIfExists('product_recipe');
+    }
+};