+++ database/seeders/ProductSeeder.php
@@ -15,6 +15,6 @@ class ProductSeeder extends Seeder
      */
     public function run(): void
     {
-        Product::factory(10)->create();
+        Product::factory(20)->create();
     }
 }